plot3D likelihood - point3.2
Guys in the code part they required a plot of joint likelihood, so basically a 3D plot. I did it but for the first data-point seems wrong if we consider the printed coordinate of the max value. Otherwise if you look to other datapoint, eg 10, everything seems all right. I really cannot understand, anyway that plot is not required in the report, so i wrote my report part and i would share my modify on the code inviting you to check if i'm wrong. Thanks guys and sorry for the long message.